working on templates selection
