+++ apps/authoring/src/main/models/projects/model-project.mock.ts
@@ -13,7 +13,8 @@ export const mockSlide = (
     lessonID: lessonID,
     moduleID: moduleID,
     name: name,
-    template: templateManifestIntro as TemplateManifest,
+    // template: templateManifestIntro as TemplateManifest,
+    template: templateManifestIntro as any,
   };
 };
