+++ apps/authoring/src/main/models/templates/assets/template-introduction/manifest.json
@@ -5,7 +5,6 @@
   },
   "meta": {
     "name": "introduction",
-    "filename": "introduction",
     "component": "Introduction"
   },
   "elements": {
