+++ apps/authoring/src/renderer/components/app/styles/comp-app.module.scss
@@ -27,3 +27,22 @@ body {
   color: var(--owl-footer-color);
   background: var(--owl-footer-bg);
 }
+
+.scrowl-template-browser {
+  display: flex;
+  justify-content: space-evenly;
+
+  &--item {
+      border: none;
+      background: none;
+      box-shadow: var(--owl-box-shadow);
+      padding: 0;
+    
+    &:hover{
+      transform: translate(-2px, -2px);
+      box-shadow: var(--owl-box-shadow-active);
+      transition: box-shadow 0.1s ease-in-out, transform 0.1s ease-in-out;
+    }
+  }
+}
+
