+++ apps/authoring/src/renderer/components/modals/index.tsx
@@ -1,2 +1,2 @@
 export * from './projects/comp-modal-project-explorer';
-export * from './templates/comp-modal-template-explorer';
+export * from '../../pages/editor/elements/modals/templates/comp-modal-template-explorer';
