+++ apps/authoring/src/renderer/pages/editor/elements/pane-details/editor-pane-details.module.scss
@@ -452,3 +452,20 @@
 .owlui-offcanvas-header {
   font-weight: normal !important;
 }
+
+.scrowl-template-browser {
+  display: flex;
+  justify-content: space-evenly;
+  &--item {
+      border: none;
+      background: none;
+      box-shadow: var(--owl-box-shadow);
+      padding: 0;
+    
+    &:hover{
+      transform: translate(-2px, -2px);
+      box-shadow: var(--owl-box-shadow-active);
+      transition: box-shadow 0.1s ease-in-out, transform 0.1s ease-in-out;
+    }
+  }
+}
